rename: 게시글 api 패키지 이름 변경 (post -> board)
