+++ src/main/java/org/jullaene/walkmong_back/api/board/domain/enums/WalkingStatus.java
@@ -1,0 +1,5 @@
+package org.jullaene.walkmong_back.api.board.domain.enums;
+
+public enum WalkingStatus {
+    BEFORE, ING, AFTER;
+}
