+++ src/main/java/org/jullaene/walkmong_back/api/board/repository/PostRepository.java
@@ -1,6 +1,6 @@
-package org.jullaene.walkmong_back.api.post.repository;
+package org.jullaene.walkmong_back.api.board.repository;
 
-import org.jullaene.walkmong_back.api.post.domain.Board;
+import org.jullaene.walkmong_back.api.board.domain.Board;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
