+++ src/main/java/org/jullaene/walkmong_back/api/board/repository/ReviewRepository.java
@@ -1,6 +1,6 @@
-package org.jullaene.walkmong_back.api.post.repository;
+package org.jullaene.walkmong_back.api.board.repository;
 
-import org.jullaene.walkmong_back.api.post.domain.Review;
+import org.jullaene.walkmong_back.api.board.domain.Review;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
