+++ src/main/java/org/jullaene/walkmong_back/api/board/service/ReviewService.java
@@ -1,7 +1,7 @@
-package org.jullaene.walkmong_back.api.post.service;
+package org.jullaene.walkmong_back.api.board.service;
 
 import lombok.RequiredArgsConstructor;
-import org.jullaene.walkmong_back.api.post.repository.ReviewRepository;
+import org.jullaene.walkmong_back.api.board.repository.ReviewRepository;
 import org.springframework.stereotype.Service;
 
 @Service
