--- src/main/java/org/jullaene/walkmong_back/api/post/domain/enums/WalkingStatus.java
@@ -1,5 +1,0 @@
-package org.jullaene.walkmong_back.api.post.domain.enums;
-
-public enum WalkingStatus {
-    BEFORE, ING, AFTER;
-}
